# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AirFlex API",
    "version": "1.0.0",
    "description": "Open-source P2P Airtime & Data Exchange Marketplace backed by the Stellar Network. All protected endpoints require a Bearer JWT obtained from POST /api/v1/auth/verify-otp.",
    "contact": {
      "name": "AirFlex Core Team",
      "url": "https://github.com/arflexx/Airflex"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/v1/auth/verify-otp. Pass as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Descriptive error message"
          }
        }
      },
      "ValidationErrorItem": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "example": "phone"
          },
          "message": {
            "type": "string",
            "example": "phone is required"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        }
      },
...
      "RequestOtpBody": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "description": "E.164 or Nigerian local format (0XXXXXXXXXX). 10–15 digits.",
            "example": "+2348012345678"
          }
        }
      },
      "VerifyOtpBody": {
        "type": "object",
        "required": [
          "phone",
          "otp"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "example": "+2348012345678"
          },
          "otp": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "pattern": "^\\d{6}$",
            "example": "123456"
          }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "Signed JWT — valid for 7 days.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "type": "object",
            "required": [
              "id",
              "phone",
              "stellarPublicKey"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "phone": {
                "type": "string",
                "example": "+2348012345678"
              },
              "stellarPublicKey": {
                "type": "string",
                "example": "GABC1234..."
              }
            }
          },
          "recoveryCodes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ABC23456789ABCDE"
            },
            "description": "8 single-use backup codes. Present ONLY on the very first signup (issue #108) — never on later logins."
          }
        }
      },
      "RecoverBody": {
        "type": "object",
        "required": [
          "recoveryCode"
        ],
        "properties": {
          "recoveryCode": {
            "type": "string",
            "minLength": 16,
            "maxLength": 16,
            "pattern": "^[A-Za-z0-9]{16}$",
            "description": "A single-use backup code issued at signup (issue #108).",
            "example": "ABC23456789ABCDE"
          }
        }
      },
      "RecoveryTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "message",
          "expiresIn"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "One-time JWT (15 min) scoped to recovery — present it to POST /api/v1/auth/recover/change-phone."
          },
          "message": {
            "type": "string"
          },
          "expiresIn": {
            "type": "string",
            "example": "15m"
          }
        }
      },
      "ChangePhoneBody": {
        "type": "object",
        "required": [
          "token",
          "newPhone"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "The one-time recovery JWT from POST /api/v1/auth/recover."
          },
          "newPhone": {
            "type": "string",
            "description": "E.164 or Nigerian local format (0XXXXXXXXXX). 10–15 digits.",
            "example": "+2348098765432"
          }
        }
      },
      "RecoveryStatusResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "remaining"
            ],
            "properties": {
              "remaining": {
                "type": "integer",
                "description": "How many backup codes remain unused (issue #108).",
                "example": 8
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid Bearer JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Missing or invalid Authorization header"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Trade offer not found"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Request body failed schema validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Too many requests. Please slow down."
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/request-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request OTP",
        "description": "Sends a 6-digit OTP to the provided phone number via Termii SMS. Creates a new user record on first call (upsert).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP sent successfully"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Termii SMS service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to send OTP. Please try again."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify OTP and obtain JWT",
        "description": "Verifies the OTP against Termii. On success issues a 7-day JWT and provisions a Stellar wallet for new users.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified — JWT issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Termii verification service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/recover": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Recover account with a backup code",
        "description": "Redeems a single-use backup code (issue #108) and returns a short-lived one-time token that can be used to register a new phone number. 5 failed attempts from the same IP within 1 hour trigger a 429 lockout.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code accepted — one-time recovery token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or already-used recovery code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/auth/recover/change-phone": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new phone number after recovery",
        "description": "Presents the one-time recovery token from POST /api/v1/auth/recover and moves the account to a new phone number (issue #108). Returns a fresh session JWT so the user is signed in immediately.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePhoneBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phone number updated — session JWT issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Recovery token is invalid, expired, or not recovery-scoped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "The new phone number is already registered to another account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/auth/recovery-codes/status": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Check remaining recovery codes",
        "description": "Returns how many backup codes remain unused for the authenticated user (issue #108). The codes themselves are never revealed.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Remaining code count.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
}